refactor: package structure to move Card, Deck, and Player classes to hwr.oop.core
