--- src/main/kotlin/hwr/oop/KotlinExample.kt
@@ -1,7 +1,0 @@
-package hwr.oop
-
-class KotlinExample {
-  fun sayHello(): String {
-    return "Hello World!"
-  }
-}
