--- src/main/kotlin/hwr/oop/Main.kt
@@ -1,6 +1,0 @@
-package hwr.oop
-
-fun main() {
-  val answer = KotlinExample().sayHello()
-  println(answer)
-}
