+++ src/main/kotlin/hwr/oop/core/Card.kt
@@ -1,3 +1,3 @@
-package hwr.oop
+package hwr.oop.core
 
 data class Card(val suit: String, val rank: String)
