+++ src/main/kotlin/hwr/oop/core/Player.kt
@@ -1,4 +1,4 @@
-package hwr.oop
+package hwr.oop.core
 
 class Player(
     val name: String,
