--- src/test/kotlin/hwr/oop/KotlinExampleTest.kt
@@ -1,14 +1,0 @@
-package hwr.oop
-
-import io.kotest.core.spec.style.AnnotationSpec
-import org.assertj.core.api.Assertions.assertThat
-
-// TODO Delete this placeholder test.
-class KotlinExampleTest : AnnotationSpec() {
-  @Test
-  fun `example returns hello world`() {
-    val example = KotlinExample()
-    val result = example.sayHello()
-    assertThat(result).isEqualTo("Hello World!")
-  }
-}
